github-graphql-query.sh: fix reviewed-by tags on Ubuntu systems
An additional "-e" was added to the signed-off-by lines on
systems with non-bash default shell (e.g. default dash in Ubuntu).
The echo command in dash does not know about "-e", so it ends
up in output (but escape sequences are handled by default). In
bash, the "-e" option has to be added to handle the newfeed
properly.

Fix the issue by using /bin/echo directly.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>